requirements.txt
ruamel.yaml==0.18.10
websockets==14.2
openai==1.61.0
loguru==0.7.3